fighting to get a way to be able to feed zips to a repackaging using the same path list constructs as can feed into tasks like javac.
ant is fighting me unbelievably hard here.  a feature for exactly this was added in 1.8, but the implementation for it is so poor that it's literally unusable.

i'm still not aware of any other java build tools that are free from bananas download-plugins-at-runtime behavior, but make no mistake, if I find one, this sort of shitslog in ant will make me switch in a hot second; and if i don't find one, i'm eventually going to boil over with rage and make one, because This Should Not Be Hard.