reform this into something that can actually be used via parameters.
could also take this one step further and accept a whole element full of paths, join those internally, and then roll into strings and etc.  but i have a hard time coming up with a situation when that's very helpful, and if it is, one can do that on the outside.  so, we only accept a refid for now.